Support for specifiying an option multiple times
This allows an option to be specified multiple times on the command-line
and have the option values stored in a list in the order that they are
specified on the command-line. Just pass `{}` as the default value to an
option, and the option key will be populated with a list of all values
specified for that key.