Modify `cli:parse` to take arg as input
This modifies the `cli:parse` function to take a table of command line
arguments as the first parameter to the function. If an argument list is
not is not passed in, it will default to the global `arg`.

This allows `cli:parse` to be more testable without having to clobber
the global `arg` variable. It also enables an application to pass
command line arguments to `cli:parse` without clobbering `_G.arg`.