Update callback to support error messages
Add an extra argument `opt` to the callback signature, which is the
option as specified on the command-line (e.g. '-k' or '--expanded-key').
The callback function just needs to return `nil, err_message` and parse
will stop parsing anymore options and automatically print the error
message if `noprint` is `false`.