Accept run-time configuration from files
This patch brings the ability to override default values for options and
flags from a config file. Supported formats are JSON, YAML, and INI, as
well as low-level helpers for implementing a custom loader.

The client is expected to have the necessary backend modules installed
for these to work. They are available from luarocks. The backends are:

- JSON  => the "dkjson" rock
- YAML  => the "yaml" rock
- INI   => the "inifile" rock