New take on loading config files
A new hook cli:load_config_from_file() lets us re-define defaults from
config files. Options that could not be mapped to CLI arguments will be
tracked in a secondary storage which is accessible through another hook,
cli:get_arbitrary_options(), and it is up to the application to handle
these options as it sees fit.

The loader can accept a specific format, or it will try to infer it from
the file extension.