fix!: Unbound splat arguments
BREAKING CHANGE: The default maxcount value for SPLAT arguments is now
unlimited. The previous default was 1. That was unexpected for many
folks and arguably wrong, but fixing it is a major breaking change and
consumers will need to adjust, hence the major version bump. The old
behaviour can be achieved by setting the maxcount to 1 again:

    cli:splat('MY_SPLAT', 'Description', nil, 1)

Additionally a limit of 999 has been resolved so unlimited arguments can
be accepted.